requirements.txt
django
django-crispy-forms
djangorestframework
drf-nested-routers